added REST API Wrappers

+++ applications/content/src/main/java/org/apache/ofbiz/content/content/ContentSearchEvents.java
@@ -27,12 +27,12 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSession;
 
 import org.apache.ofbiz.base.util.UtilValidate;
-import org.apache.ofbiz.content.content.ContentSearch.ContentSearchConstraint;
-import org.apache.ofbiz.content.content.ContentSearch.ContentSearchContext;
-import org.apache.ofbiz.content.content.ContentSearch.ResultSortOrder;
-import org.apache.ofbiz.content.content.ContentSearchSession.ContentSearchOptions;
 import org.apache.ofbiz.entity.Delegator;
 import org.apache.ofbiz.webapp.stats.VisitHandler;
+import org.apache.ofbiz.content.content.ContentSearch.ResultSortOrder;
+import org.apache.ofbiz.content.content.ContentSearch.ContentSearchConstraint;
+import org.apache.ofbiz.content.content.ContentSearch.ContentSearchContext;
+import org.apache.ofbiz.content.content.ContentSearchSession.ContentSearchOptions;
 
 
 public class ContentSearchEvents {
@@ -52,16 +52,16 @@ public class ContentSearchEvents {
         ContentSearchOptions contentSearchOptions = ContentSearchSession.getContentSearchOptions(session);
 
         Integer viewIndexInteger = contentSearchOptions.getViewIndex();
-        if (viewIndexInteger != null) viewIndex = viewIndexInteger;
+        if (viewIndexInteger != null) viewIndex = viewIndexInteger.intValue();
         Integer viewSizeInteger = contentSearchOptions.getViewSize();
-        if (viewSizeInteger != null) viewSize = viewSizeInteger;
+        if (viewSizeInteger != null) viewSize = viewSizeInteger.intValue();
 
         lowIndex = viewIndex * viewSize;
         highIndex = (viewIndex + 1) * viewSize;
 
         // setup resultOffset and maxResults, noting that resultOffset is 1 based, not zero based as these numbers
-        Integer resultOffset = lowIndex + 1;
-        Integer maxResults = viewSize;
+        Integer resultOffset = Integer.valueOf(lowIndex + 1);
+        Integer maxResults = Integer.valueOf(viewSize);
 
         // ========== Do the actual search
         ArrayList<String> contentIds = null;
@@ -85,7 +85,7 @@ public class ContentSearchEvents {
 
             Integer totalResults = contentSearchContext.getTotalResults();
             if (totalResults != null) {
-                listSize = totalResults;
+                listSize = totalResults.intValue();
             }
         }
 
@@ -101,11 +101,11 @@ public class ContentSearchEvents {
         Map<String, Object> result = new HashMap<String, Object>();
 
         result.put("contentIds", contentIds);
-        result.put("viewIndex", viewIndex);
-        result.put("viewSize", viewSize);
-        result.put("listSize", listSize);
-        result.put("lowIndex", lowIndex);
-        result.put("highIndex", highIndex);
+        result.put("viewIndex", Integer.valueOf(viewIndex));
+        result.put("viewSize", Integer.valueOf(viewSize));
+        result.put("listSize", Integer.valueOf(listSize));
+        result.put("lowIndex", Integer.valueOf(lowIndex));
+        result.put("highIndex", Integer.valueOf(highIndex));
         result.put("searchConstraintStrings", searchConstraintStrings);
         result.put("searchSortOrderString", searchSortOrderString);
