+++ framework/minilang/src/main/java/org/apache/ofbiz/minilang/method/otherops/Log.java
@@ -31,7 +31,7 @@ import org.w3c.dom.Element;
 /**
  * Implements the &lt;log&gt; element.
  * 
- * @see <a href="https://cwiki.apache.org/confluence/display/OFBIZ/Mini+Language+-+minilang+-+simple-method+-+Reference">Mini-language Reference</a>
+ * @see <a href="https://cwiki.apache.org/confluence/display/OFBADMIN/Mini-language+Reference#Mini-languageReference-{{%3Clog%3E}}">Mini-language Reference</a>
  */
 public final class Log extends MethodOperation {
 
@@ -60,7 +60,7 @@ public final class Log extends MethodOperation {
             MiniLangValidate.handleError("Invalid level attribute", simpleMethod, element);
             this.level = Debug.INFO;
         } else {
-            this.level = levelInt;
+            this.level = levelInt.intValue();
         }
     }
