+++ framework/widget/src/main/java/org/apache/ofbiz/widget/model/ModelWidget.java
@@ -63,8 +63,8 @@ public abstract class ModelWidget implements Serializable {
     protected ModelWidget(Element widgetElement) {
         this.name = widgetElement.getAttribute("name");
         this.systemId = (String) widgetElement.getUserData("systemId");
-        this.startColumn = (Integer) widgetElement.getUserData("startColumn");
-        this.startLine = (Integer) widgetElement.getUserData("startLine");
+        this.startColumn = ((Integer) widgetElement.getUserData("startColumn")).intValue();
+        this.startLine = ((Integer) widgetElement.getUserData("startLine")).intValue();
     }
 
     public abstract void accept(ModelWidgetVisitor visitor) throws Exception;
@@ -130,7 +130,7 @@ public abstract class ModelWidget implements Serializable {
      * rendering context. If <code>widget.verbose</code> is set to <code>false</code> in the
      * <code>widget.properties</code> file, then that setting will override all other settings and
      * disable all widget boundary comments.
-     *
+     * 
      * @param context Optional context Map
      */
     public static boolean widgetBoundaryCommentsEnabled(Map<String, ? extends Object> context) {
@@ -139,7 +139,7 @@ public abstract class ModelWidget implements Serializable {
             String str = (String) context.get(enableBoundaryCommentsParam);
             if (str != null) {
                 result = "true".equals(str);
-            } else {
+            } else{
                 Map<String, ? extends Object> parameters = UtilGenerics.checkMap(context.get("parameters"));
                 if (parameters != null) {
                     str = (String) parameters.get(enableBoundaryCommentsParam);
